+++ test/IntegrationTests/SmokeTests.cs
@@ -497,6 +497,11 @@ public class SmokeTests : TestHelper
         resourceExpector.Expect("telemetry.sdk.version", typeof(OpenTelemetry.Resources.Resource).Assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()!.InformationalVersion.Split('+')[0]);
         resourceExpector.Expect("telemetry.distro.name", "opentelemetry-dotnet-instrumentation");
         resourceExpector.Expect("telemetry.distro.version", typeof(OpenTelemetry.AutoInstrumentation.Constants).Assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()!.InformationalVersion.Split('+')[0]);
+#if NETFRAMEWORK
+        resourceExpector.Expect("process.runtime.name", ".NET Framework");
+#else
+        resourceExpector.Expect("process.runtime.name", ".NET");
+#endif
     }
 
     private void VerifyTestApplicationInstrumented(TestAppStartupMode startupMode = TestAppStartupMode.Auto)